initial cleanup of all data reader / layer remove unused reader
