+++ ecal/core/src/readwrite/ecal_reader.h
@@ -101,8 +101,8 @@ namespace eCAL
     size_t AddSample(const std::string& tid_, const char* payload_, size_t size_, long long id_, long long clock_, long long time_, size_t hash_, eCAL::pb::eTLayerType layer_);
 
   protected:
-    void StartDataLayers();
-    void StopDataLayers();
+    void SubscribeToLayers();
+    void UnsubscribeFromLayers();
 
     bool DoRegister(const bool force_);
     void SetConnected(bool state_);
