+++ ecal/core/src/readwrite/ecal_reader_layer.h
@@ -31,7 +31,7 @@
 namespace eCAL
 {
   // ecal data layer specific parameters
-  // will be transmitted between writer and reader
+  // transmitted from a writer to a reader
   struct SReaderLayerPar
   {
     std::string host_name;
@@ -50,11 +50,17 @@ namespace eCAL
 
     virtual ~CReaderLayer() = default;
 
+    // initialize layer
+    // will be called one time on eCAL initialization
     virtual void Initialize() = 0;
 
+    // activate / create a specific subscription
     virtual void AddSubscription(const std::string& host_name_, const std::string& topic_name_, const std::string& topic_id_, QOS::SReaderQOS qos_) = 0;
+
+    // deactivate / destroy a specific subscription
     virtual void RemSubscription(const std::string& host_name_, const std::string& topic_name_, const std::string& topic_id_) = 0;
 
+    // connection paramter from writer side
     virtual void SetConnectionParameter(SReaderLayerPar& par_) = 0;
 
     static std::shared_ptr<T> Get()
