+++ ecal/core/src/readwrite/ecal_reader_shm.h
@@ -18,7 +18,7 @@
 */
 
 /**
- * @brief  memory file data reader
+ * @brief  shared memory layer
 **/
 
 #pragma once
@@ -30,34 +30,28 @@
 
 namespace eCAL
 {
-  // ecal shared memory reader
-  class CDataReaderSHM
-  {
-  public:
-    CDataReaderSHM() = default;
-    ~CDataReaderSHM() = default;
-  };
+  ////////////////
+  // READER
+  ////////////////
+  // class CDataReaderSHM
+  // {
+  // public:
+    // CDataReaderSHM() = default;
+    // ~CDataReaderSHM() = default;
+  // };
 
-  // ecal shared memory layer
+  ////////////////
+  // LAYER
+  ////////////////
   class CSHMLayer : public CReaderLayer<CSHMLayer>
   {
   public:
-    CSHMLayer() {};
-    ~CSHMLayer()
-    {
-    };
+    CSHMLayer() {}
+    ~CSHMLayer() {}
 
-    void Initialize()
-    {
-    }
-
-    void AddSubscription(const std::string& /*host_name_*/, const std::string& /*topic_name_*/, const std::string& /*topic_id_*/, QOS::SReaderQOS /*qos_*/)
-    {
-    }
-
-    void RemSubscription(const std::string& /*host_name_*/, const std::string& /*topic_name_*/, const std::string& /*topic_id_*/)
-    {
-    }
+    void Initialize() {}
+    void AddSubscription(const std::string& /*host_name_*/, const std::string& /*topic_name_*/, const std::string& /*topic_id_*/, QOS::SReaderQOS /*qos_*/) {}
+    void RemSubscription(const std::string& /*host_name_*/, const std::string& /*topic_name_*/, const std::string& /*topic_id_*/) {}
 
     void SetConnectionParameter(SReaderLayerPar& par_);
   };
